# Make WrappedTransform work with --class-log-level